Updating our controller

+++ src/controller.py
@@ -149,7 +149,7 @@ def get_todos(current_user):
   if not current_user.admin:
     return jsonify({'msg': 'Cannot perform that function!'})
 
-  todos=Todo.query.all()
+  todos=Todo.query.filter_by(user_id=current_user.id).all()
   result = todos_schema.dump(todos)
   return jsonify(result)
 
@@ -159,7 +159,10 @@ def get_todo(current_user, id):
   if not current_user.admin:
     return jsonify({'msg': 'Cannot perform that function!'})
 
-  todo=Todo.query.get(id)
+  todo=Todo.query.filter_by(id=id, user_id=current_user.id).first()
+  if not todo:
+    return jsonify({'msg': 'No todo found!'})
+    
   return todo_schema.jsonify(todo)
 
 @app.route('/v1/todos', methods=['POST'])
@@ -170,7 +173,7 @@ def create_todo(current_user):
 
   content=request.json['content']
   done=False
-  user_id=request.json['user_id']  
+  user_id=current_user.id  
 
   new_todo=Todo(content,done,user_id)
 
@@ -185,11 +188,14 @@ def update_todo(current_user, id):
   if not current_user.admin:
     return jsonify({'msg': 'Cannot perform that function!'})
 
-  todo=Todo.query.get(id)
+  todo=Todo.query.filter_by(id=id, user_id=current_user.id).first()
+
+  if not todo:
+    return jsonify({'msg': 'No todo found!'})
 
   content=request.json['content']
   done=request.json['done']
-  user_id=request.json['user_id']
+  user_id=current_user.id
 
   todo.content=content
   todo.done=done
@@ -205,7 +211,10 @@ def delete_todo(current_user, id):
   if not current_user.admin:
     return jsonify({'msg': 'Cannot perform that function!'})
 
-  todo=Todo.query.get(id)
+  todo=Todo.query.filter_by(id=id, user_id=current_user.id).first()
+
+  if not todo:
+    return jsonify({'msg': 'No todo found!'})
 
   db.session.delete(todo)
   db.session.commit()
